+++ spec/hunspell-asm/util/logger-spec.ts
@@ -1,15 +1,7 @@
-//tslint:disable:no-require-imports
 import { expect } from 'chai';
-import loggerType = require('../../../src/util/logger');
+import { enableLogger, log } from '../../../src/util/logger';
 
 describe('logger', () => {
-  let log: typeof loggerType.log;
-  let enableLogger: typeof loggerType.enableLogger;
-
-  beforeEach(() => {
-    ({ log, enableLogger } = require('../../../src/util/logger'));
-  });
-
   it('should do nothing by default', () => {
     expect(() => log('')).to.not.throw();
   });
@@ -24,6 +16,6 @@ describe('logger', () => {
     log(message, value);
 
     expect(mock.mock.calls).to.have.lengthOf(1);
-    expect(mock.mock.calls[0]).to.deep.equal([message, value]);
+    expect(mock.mock.calls[0]).to.deep.equal([`hunspell::${message}`, value]);
   });
 });